requirements.txt
spacy>=3.7.0
PyMuPDF>=1.23.0
pdfminer.six>=20221105
streamlit>=1.28.0
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0